Moving env helpers into separate file.
