+++ resources/assets/components/CampaignBanner/CampaignBanner.js
@@ -5,13 +5,8 @@ import classnames from 'classnames';
 import { useQuery } from '@apollo/react-hooks';
 import React, { useState, useEffect } from 'react';
 
-import {
-  isScholarshipAffiliateReferral,
-  getScholarshipAffiliateLabel,
-  isCurrentPathInPaths,
-  siteConfig,
-} from '../../helpers';
 import Modal from '../utilities/Modal/Modal';
+import { siteConfig } from '../../helpers/env';
 import ContentfulEntry from '../ContentfulEntry';
 import partnerScholarshipQuizPaths from './config';
 import Spinner from '../artifacts/Spinner/Spinner';
@@ -28,6 +23,11 @@ import AffiliatePromotion from '../utilities/AffiliatePromotion/AffiliatePromoti
 import ScholarshipInfoBlock from '../blocks/ScholarshipInfoBlock/ScholarshipInfoBlock';
 import CampaignSignupFormContainer from '../CampaignSignupForm/CampaignSignupFormContainer';
 import AffiliateOptInToggleContainer from '../AffiliateOptInToggle/AffiliateOptInToggleContainer';
+import {
+  isScholarshipAffiliateReferral,
+  getScholarshipAffiliateLabel,
+  isCurrentPathInPaths,
+} from '../../helpers';
 import ScholarshipReferralVoterRegistrationBlock from '../blocks/ScholarshipReferralVoterRegistrationBlock/ScholarshipReferralVoterRegistrationBlock';
 
 const CAMPAIGN_BANNER_QUERY = gql`
