+++ resources/assets/components/CampaignSignupForm/CampaignSignupForm.js
@@ -3,14 +3,10 @@ import React, { useState } from 'react';
 
 import Card from '../utilities/Card/Card';
 import { getUtms } from '../../helpers/utm';
+import { siteConfig } from '../../helpers/env';
 import GroupFinder from './GroupFinder/GroupFinder';
 import PrimaryButton from '../utilities/Button/PrimaryButton';
-import {
-  isCampaignClosed,
-  query,
-  siteConfig,
-  withoutNulls,
-} from '../../helpers';
+import { isCampaignClosed, query, withoutNulls } from '../../helpers';
 import { EVENT_CATEGORIES, trackAnalyticsEvent } from '../../helpers/analytics';
 
 const CampaignSignupForm = props => {
