+++ resources/assets/components/actions/PetitionSubmissionAction/PetitionSubmissionAction.js
@@ -7,7 +7,7 @@ import { Query as ApolloQuery } from 'react-apollo';
 
 import PostForm from '../PostForm';
 import Card from '../../utilities/Card/Card';
-import { featureFlag } from '../../../helpers';
+import { featureFlag } from '../../../helpers/env';
 import PostCreatedModal from '../PostCreatedModal';
 import ActionInformation from '../ActionInformation';
 import FormValidation from '../../utilities/Form/FormValidation';
