+++ resources/assets/components/pages/QuizResultPage/QuizResultPage.js
@@ -7,8 +7,8 @@ import ErrorPage from '../ErrorPage';
 import triangle from './triangle.svg';
 import gqlVariables from './config';
 import NotFoundPage from '../NotFoundPage';
-import { isDevEnvironment } from '../../../helpers';
 import Placeholder from '../../utilities/Placeholder';
+import { isDevEnvironment } from '../../../helpers/env';
 import SiteFooter from '../../utilities/SiteFooter/SiteFooter';
 import TextContent from '../../utilities/TextContent/TextContent';
 import { LinkBlockFragment } from '../../actions/LinkAction/LinkAction';
