+++ resources/assets/components/pages/ReferralPage/Beta/BetaPage.js
@@ -3,10 +3,11 @@ import gql from 'graphql-tag';
 
 import Query from '../../../Query';
 import ErrorPage from '../../ErrorPage';
+import { query } from '../../../../helpers';
 import CampaignLink from './BetaPageCampaignLink';
 import GiftCardHandImage from './gift-card-hand.svg';
 import ArticleHeader from '../../../utilities/ArticleHeader';
-import { featureFlag, query, siteConfig } from '../../../../helpers';
+import { siteConfig, featureFlag } from '../../../../helpers/env';
 import { getReferralCampaignId } from '../../../../helpers/refer-friends';
 
 export const REFERRAL_USER_QUERY = gql`
