+++ resources/assets/components/pages/VoterRegistrationDrivePage/VoterRegistrationDrivePage.js
@@ -4,10 +4,11 @@ import { useQuery } from 'react-apollo';
 
 import ErrorPage from '../ErrorPage';
 import { gqlVariables } from './config';
+import { query } from '../../../helpers';
 import NotFoundPage from '../NotFoundPage';
 import Modal from '../../utilities/Modal/Modal';
 import Placeholder from '../../utilities/Placeholder';
-import { isDevEnvironment, query } from '../../../helpers';
+import { isDevEnvironment } from '../../../helpers/env';
 import SiteFooter from '../../utilities/SiteFooter/SiteFooter';
 import PrimaryButton from '../../utilities/Button/PrimaryButton';
 import VoterRegistrationDrivePageBanner from './VoterRegistrationDrivePageBanner';
