+++ resources/assets/components/utilities/ContentfulAsset/ContentfulAsset.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { Query } from 'react-apollo';
 
 import LazyImage from '../LazyImage';
-import { env } from '../../../helpers';
+import { env } from '../../../helpers/env';
 import Spinner from '../../artifacts/Spinner/Spinner';
 import ErrorBlock from '../../blocks/ErrorBlock/ErrorBlock';
