+++ resources/assets/components/utilities/PopoverDispatcher/CtaPopover/CtaPopoverEmailForm.js
@@ -3,9 +3,10 @@ import React, { useState } from 'react';
 import { get, isString, first } from 'lodash';
 import { RestApiClient } from '@dosomething/gateway';
 
+import { env } from '../../../../helpers/env';
+import { report } from '../../../../helpers/index';
 import { tabularLog } from '../../../../helpers/api';
 import PrimaryButton from '../../Button/PrimaryButton';
-import { env, report } from '../../../../helpers/index';
 import {
   EVENT_CATEGORIES,
   trackAnalyticsEvent,
