+++ resources/assets/helpers/facebook.js
@@ -1,6 +1,7 @@
 /* global window, document */
 
-import { env, makeUrl, openDialog } from '.';
+import { env } from './env';
+import { makeUrl, openDialog } from '.';
 import { EVENT_CATEGORIES, trackAnalyticsEvent } from './analytics';
 
 /**
