+++ resources/assets/helpers/index.js
@@ -204,17 +204,6 @@ export function isAuthenticated() {
 }
 
 /**
- * Check to see if running in a development enviroment.
- *
- * @return {Boolean}
- */
-export function isDevEnvironment() {
-  return ['local', 'development'].includes(
-    get(window.ENV, 'APP_ENV', 'production'),
-  );
-}
-
-/**
  * Check to see if user is a DS.org staffer.
  * @todo move to helpers/auth.js
  *
@@ -489,38 +478,6 @@ export function isWithinMinutes(date, minutes = 2) {
  */
 export function findById(array, compareId) {
   return array.find(({ id }) => id === compareId);
-}
-
-/**
- * Get a value from the window.ENV
- *
- * @param  {String} key
- * @return {String}
- */
-export function env(key, defaultVal) {
-  return get(window.ENV, key, defaultVal);
-}
-
-/**
- * Get feature flag value from window.ENV
- *
- * @param  {String} feature
- * @param  {*}      defaultVal
- * @return {*}
- */
-export function featureFlag(feature, defaultVal) {
-  return get(env('FEATURE_FLAGS'), feature, defaultVal);
-}
-
-/**
- * Get site config value from window.ENV
- *
- * @param  {String} config
- * @param  {*}      defaultVal
- * @return {*}
- */
-export function siteConfig(config, defaultVal) {
-  return get(env('SITE'), config, defaultVal);
 }
 
 /**
